SYS-3773 Separate request actions from transactions
