+++ pallets/eth-bridge/src/tx.rs
@@ -3,41 +3,8 @@ use crate::{offence::create_and_report_corroboration_offence, util::bound_params
 use frame_support::BoundedVec;
 use sp_core::Get;
 
-pub fn add_new_request<T: Config>(
-    function_name: &[u8],
-    params: &[(Vec<u8>, Vec<u8>)],
-) -> Result<EthereumId, Error<T>> {
-    let function_name_string =
-        String::from_utf8(function_name.to_vec()).map_err(|_| Error::<T>::FunctionNameError)?;
-    if function_name_string.is_empty() {
-        return Err(Error::<T>::EmptyFunctionName)
-    }
-
-    let tx_id = use_next_tx_id::<T>();
-
-    let tx_request = Request {
-        tx_id,
-        function_name: BoundedVec::<u8, FunctionLimit>::try_from(function_name.to_vec())
-            .map_err(|_| Error::<T>::ExceedsFunctionNameLimit)?,
-        params: bound_params(&params.to_vec())?,
-    };
-
-    if ActiveRequest::<T>::get().is_some() {
-        queue_tx_request(tx_request)?;
-    } else {
-        set_up_active_tx(tx_request)?;
-    }
-
-    Ok(tx_id)
-}
-
 pub fn is_active<T: Config>(tx_id: EthereumId) -> bool {
     ActiveRequest::<T>::get().map_or(false, |active_tx| active_tx.id == tx_id)
-}
-
-fn replay_transaction<T: Config>(mut tx: ActiveTransactionData<T>) -> Result<(), Error<T>> {
-    tx.request_data.tx_id = use_next_tx_id::<T>();
-    Ok(set_up_active_tx(tx.request_data)?)
 }
 
 fn complete_transaction<T: Config>(
@@ -77,7 +44,7 @@ fn complete_transaction<T: Config>(
     // Write the tx data to permanent storage:
     SettledTransactions::<T>::insert(tx.id, tx.data);
 
-    if let Some(tx_request) = dequeue_tx_request::<T>() {
+    if let Some(tx_request) = request::dequeue_tx_request::<T>() {
         set_up_active_tx(tx_request)?;
     } else {
         ActiveRequest::<T>::kill();
@@ -94,46 +61,13 @@ pub fn finalize_state<T: Config>(
     // replay transaction
     if !success && util::has_enough_corroborations::<T>(tx.invalid_tx_hash_corroborations.len()) {
         // raise an offence on the "sender" because the tx_hash they provided was invalid
-        return Ok(replay_transaction(tx)?)
+        return Ok(request::replay_transaction(tx)?)
     }
 
     Ok(complete_transaction::<T>(tx, success)?)
 }
 
-fn queue_tx_request<T: Config>(tx_request: Request) -> Result<(), Error<T>> {
-    RequestQueue::<T>::mutate(|maybe_queue| {
-        let mut queue: Vec<_> = maybe_queue.clone().unwrap_or_else(Default::default).into();
-
-        if queue.len() < T::MaxQueuedTxRequests::get() as usize {
-            queue.push(tx_request);
-            let bounded_queue =
-                BoundedVec::try_from(queue).expect("Size known to be in bounds here");
-            *maybe_queue = Some(bounded_queue);
-            Ok(())
-        } else {
-            Err(Error::<T>::TxRequestQueueFull)
-        }
-    })
-}
-
-fn dequeue_tx_request<T: Config>() -> Option<Request> {
-    let mut queue = <RequestQueue<T>>::take();
-
-    let next_tx_request = match &mut queue {
-        Some(q) if !q.is_empty() => Some(q.remove(0)),
-        _ => None,
-    };
-
-    if let Some(q) = &queue {
-        if !q.is_empty() {
-            RequestQueue::<T>::put(q);
-        }
-    }
-
-    next_tx_request
-}
-
-fn set_up_active_tx<T: Config>(tx_request: Request) -> Result<(), Error<T>> {
+pub fn set_up_active_tx<T: Config>(tx_request: Request) -> Result<(), Error<T>> {
     let expiry = util::time_now::<T>() + EthTxLifetimeSecs::<T>::get();
     let data = eth::create_tx_data(&tx_request, expiry)?;
     let msg_hash = eth::generate_msg_hash(&data)?;
@@ -153,10 +87,4 @@ fn set_up_active_tx<T: Config>(tx_request: Request) -> Result<(), Error<T>> {
     });
 
     Ok(())
-}
-
-fn use_next_tx_id<T: Config>() -> EthereumId {
-    let tx_id = NextTxId::<T>::get();
-    NextTxId::<T>::put(tx_id + 1);
-    tx_id
 }
