+++ pallets/eth-bridge/src/types.rs
@@ -1,0 +1,34 @@
+use crate::*;
+
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug, Default, TypeInfo, MaxEncodedLen)]
+    pub struct Request {
+        pub tx_id: EthereumId,
+        pub function_name: BoundedVec<u8, FunctionLimit>,
+        pub params:
+            BoundedVec<(BoundedVec<u8, TypeLimit>, BoundedVec<u8, ValueLimit>), ParamsLimit>,
+    }
+
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug, Default, TypeInfo, MaxEncodedLen)]
+pub struct TransactionData<T: Config> {
+    pub function_name: BoundedVec<u8, FunctionLimit>,
+    pub params:
+        BoundedVec<(BoundedVec<u8, TypeLimit>, BoundedVec<u8, ValueLimit>), ParamsLimit>,
+    pub sender: T::AccountId,
+    pub eth_tx_hash: H256,
+    pub tx_succeeded: bool,
+}
+
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug, Default, TypeInfo, MaxEncodedLen)]
+pub struct ActiveTransactionData<T: Config> {
+    pub id: EthereumId,
+    pub request_data: Request,
+    pub data: TransactionData<T>,
+    pub expiry: u64,
+    pub msg_hash: H256,
+    pub last_updated: T::BlockNumber,
+    pub confirmations: BoundedVec<ecdsa::Signature, ConfirmationsLimit>,
+    pub success_corroborations: BoundedVec<T::AccountId, ConfirmationsLimit>,
+    pub failure_corroborations: BoundedVec<T::AccountId, ConfirmationsLimit>,
+    pub valid_tx_hash_corroborations: BoundedVec<T::AccountId, ConfirmationsLimit>,
+    pub invalid_tx_hash_corroborations: BoundedVec<T::AccountId, ConfirmationsLimit>,
+}
